feat(prompt): support custom Copilot instructions file
Add support for including custom Copilot instructions from
.github/copilot-instructions.md in the system prompt. Refactor prompt
resolution logic into a new prompt.lua module, and update function and
tool resolution to use this module. Also simplify diff hunk application
fallback logic.
