+++ lua/CopilotChat/instructions/custom_instructions.lua
@@ -1,0 +1,6 @@
+return [[
+<customInstructions>
+Custom instructions from user's `{FILENAME}`:
+{CONTENT}
+</customInstructions>
+]]
